master: added packetqueue files
packet_queue.h provides definitions for the packet_queue functions,
making a suitable FIFO queue implementation possible. 

new_packet_queue and free_packet_queue serve as constructors/destructors
respectively so that the initial setup of the queues can be controlled.

the four (acting) queue functions, add_packet, poll_next, has_pending
and get_size provide the main functionality of the queues. the name are
quite self explanatory.

additionally, synchronisation mechanisms have been added to this
interface to allow multithreaded applications to access and modify the
queue structures such that synchronisation issues cannot be present.
each of the acting functions, should make appropriate calls to lock and
unlock with the correct mutex for the current queue before acting on
them to ensure this.

conditional logic is added for global control of the threading
mechanisms, this may, however, be exchanged in the future for queue
specific thread support.

in the future, this interface may be changed to support generic queues,
i.e. void * will be accepted instead of the specific packet_t * that
this current queue supports.

packetqueue.c contains the relevant implementations of the packet_queue
functions.