master: added packethandler
the base of the server packethandling mechanism has been added in the
files: packethandler.h and packethandler.c

packethandler.h defines a server_config struct which is passed on to the
pthread server start function, void *handler_start(void *). this struct
contains three user defined fields: wsadata (from WinSock
initialisation), addr which holds the server's address, and
exit_callback, which will be called when the server handling thread is
about to exit. more callback functionality may be added later.

the field is_live refers to whether the server handler is running for a
given server_config. is_error is a boolean value indicating whether a
(fatal) error occured and if so, a more specific, null terminated
message is contained in error_msg. WSAGetLastError() may also be of use;
this is declared in WinSock.

the internal struct connection holds references relevant to a client
connection, most notably, the socket, buffers and queues for sending and
receiving data. only send_queue and recv_queue are intended to be access
by client applications.

packets pushed onto the send_queue are essentially scheduled for
transmission by the server and those in the recv_queue are pending
processing by the application code.

the two functions, dissolve_connection and create_connection are
intended to be used by the server handler only. in the case of
is_live_connection, the application may query this, however, the
implementation simply consists of a null pointer check.

packethandler.c contains the (unfinished) implementation of the server
packet handler and relies on WinSock functions to bind and setup the
server socket. connections are handled using select() for now, however,
this will probably be revised in the future, as select() only allows a
specific number of clients to be connected at once. additionally, the
actual packet handling capabilities are not yet implemented.